fixed erro for demo

+++ TiposPrimitivos_Wrapper/inteiro_positivo/InteiroPositivo.java
@@ -1,8 +1,8 @@
 public class InteiroPositivo {
     
-    private int valor;
+    private float valor;
 
-    public InteiroPositivo(int valor) {
+    public InteiroPositivo(float valor) {
         setValor(valor);
     }
